Updated WerkZeug and associated dependencies.
Werkzeug < 1.0.0 uses pyOpenSSL instead of cryptography for TLS support. Upgrading
will save adding a new dependency to the project (at the cost of a lot/little pain fixing
up deprecated functionality)

It looks like the test requirements were loaded last during the dev container setup.
The test reqs will override the securedrop-app-code reqs as a result, which means
the expected versions will not be in place for testing. As a first attempt at a fix,
the order of `pip install`s has been changed so securedrop-app-code has the last
word, but a better fix would be to generate a single requirements file for dev containing
everything needed.